Adding the use of components and changing the burn calories calculator to be used as a component, also uninstall tailwind and adding variables to the CSS
