+++ gym-app/package.json
@@ -9,14 +9,11 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "autoprefixer": "10.4.14",
     "dotenv": "^16.0.3",
     "next": "13.4.2",
-    "postcss": "8.4.23",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-redux": "^8.0.5",
-    "redux": "^4.2.1",
-    "tailwindcss": "3.3.2"
+    "redux": "^4.2.1"
   }
 }
