+++ gym-app/src/app/layout.jsx
@@ -1,4 +1,3 @@
-import './globals.css'
 import { Inter } from 'next/font/google'
 
 const inter = Inter({ subsets: ['latin'] })
@@ -11,6 +10,7 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
+      {/* <div><p>hello</p></div> */}
       <body className={inter.className}>{children}</body>
     </html>
   )
